Add zerologon_utils/vuln_assessment.py
